refactor: separate each wallet sub components
