+++ frontend/src/components/wallet/account-dropdown/index.css
