+++ frontend/src/components/wallet/native-balance/index.tsx
@@ -3,9 +3,9 @@ import { useAccount, useBalance } from 'wagmi';
 import { useWallet, useConnection } from '@solana/wallet-adapter-react';
 import { formatEther } from 'viem';
 import { LAMPORTS_PER_SOL } from '@solana/web3.js';
-import { getNativeTokenSymbol } from '../../constants/chains/ethereum';
-import Icon, { WarningIcon } from '../common/Icon';
-import './NativeBalance.css';
+import { getNativeTokenSymbol } from '../../../constants/chains/ethereum';
+import Icon, { WarningIcon } from '../../common/Icon';
+import './index.css';
 
 export type BalanceType = 'ethereum' | 'solana';
