+++ frontend/src/components/wallet/network-switcher/ethereum.tsx
