+++ frontend/src/components/wallet/network-switcher/index.ts
