+++ frontend/src/components/wallet/network-switcher/solana.tsx
