+++ frontend/src/components/wallet/solana-wallet-trigger/index.tsx
